--- src/components/AvailableProducts/index.jsx
@@ -1,8 +1,0 @@
-import {Conteiner, ConteinerDescription } from './styles'
-const AvaliableProducts = () => {
-return(
-  <Conteiner/>
-)
-}
-
-export default AvaliableProducts
