--- src/components/AvailableProducts/styles.js
@@ -1,5 +1,0 @@
-import styled from "styled-components/native";
-
-export const Conteiner = styled.View`
-`
-export const ConteinerDescription = styled.View``
