+++ src/pages/UserScreens/ProducersWithThisProduct/index.jsx
@@ -1,0 +1,9 @@
+import { Text } from "react-native"
+
+
+
+const ProducersWithThisProduct = () => {
+  return <Text> Essa é a tela que deve mostrar todos os produtores que comercializam o produto selecionado</Text>
+}
+
+export default ProducersWithThisProduct
